refactor(ai): update imports and enhance AI service functionality
- Adjusted import paths for PointsService, UserService, and OpenaiService for better module organization.
- Modified imgContentByAi and videoContentByAi methods to support optional system prompts in the options parameter.
- Exported fireflycard schemas in image.dto for improved accessibility.
- Updated VideoService to use UserKlingText2VideoRequestDto for better type consistency.
